kvm tools: ARM: set interrupt priority mask in secondary boot path
A bug in the KVM GIC init code set the priority mask to the
highest possible value, while the reset value should be zero.

Now that the kernel bug is fixed, kvm-tool must properly configure
its GIC CPU interface in order to receive the boot IPI.  Just set
the GICC_PMR register to the maximum value (0xff), and it "just works".

Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>
[will: added #define for PMR offset]
Signed-off-by: Will Deacon <will.deacon@arm.com>
Signed-off-by: Pekka Enberg <penberg@kernel.org>